refined comments

+++ scoreboard.py
@@ -7,7 +7,7 @@ SIZE = 16
 
 class Scoreboard(Turtle):
 
-    def __init__(self, score=0, x_cor=int, y_cor=int, align=str):
+    def __init__(self, score=0, x_cor=float, y_cor=float, align=str):
         super().__init__()
         self.goto(x=x_cor, y=y_cor)
         self.score = score
